win: avoid warning in StringFileWriter.SeekInvalid test
d:\src\crashpad\crashpad\util\file\string_file_writer_test.cc(367) : warning C4244: 'initializing' : conversion from 'const uint64_t' to 'const off_t', possible loss of data

R=mark@chromium.org
BUG=crashpad:1

Review URL: https://codereview.chromium.org/836413004

+++ util/file/string_file_writer_test.cc
@@ -362,9 +362,9 @@ TEST(StringFileWriter, SeekInvalid) {
   EXPECT_EQ(0, writer.Seek(0, SEEK_CUR));
   EXPECT_TRUE(writer.string().empty());
 
-  const off_t kMaxOffset =
+  const off_t kMaxOffset = static_cast<off_t>(
       std::min(implicit_cast<uint64_t>(std::numeric_limits<off_t>::max()),
-               implicit_cast<uint64_t>(std::numeric_limits<size_t>::max()));
+               implicit_cast<uint64_t>(std::numeric_limits<size_t>::max())));
 
   EXPECT_EQ(kMaxOffset, writer.Seek(kMaxOffset, SEEK_SET));
   EXPECT_EQ(kMaxOffset, writer.Seek(0, SEEK_CUR));
